+++ eventlog-live/CHANGELOG.md
@@ -3,7 +3,9 @@
 - Add `averageCounterBy` and `averageCounterByTick`.
 - Add `statsParser`.
 - Add `aggregateByTick`.
-- Add `AggregateMetrics` with suitable `Semigroup` instance.
+- Add `GHC.Eventlog.Live.Machine.Group` module for aggregation.
+- **BREAKING**: Change `processHeapProfSampleData` to yield all the metrics
+  from a single garbage collection pass at once as `HeapProfSampleData`.
 - **BREAKING**: Replace `[Attr]` with opaque `Attrs` type.
 - **BREAKING**: Log info to stderr with `--verbosity==debug`.